+++ src/test/java/grondag/exotic_matter/PolyTest.java
@@ -1,6 +1,7 @@
 package grondag.exotic_matter;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
@@ -8,8 +9,8 @@ import grondag.exotic_matter.render.FaceVertex;
 import grondag.exotic_matter.render.IMutablePolygon;
 import grondag.exotic_matter.render.Poly;
 import grondag.exotic_matter.render.QuadHelper;
+import grondag.exotic_matter.render.Vec3f;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.math.Vec3d;
 
 public class PolyTest
 {
@@ -17,27 +18,27 @@ public class PolyTest
     @Test
     public void test()
     {
-        Vec3d direction;
-        Vec3d point;
+        Vec3f direction;
+        Vec3f point;
         IMutablePolygon quad = Poly.mutable(4).setupFaceQuad(EnumFacing.UP, 0, 0, 1, 1, 0.5, EnumFacing.NORTH);
         
         
         // point on plane
-        point = new Vec3d(0.5, 0.5, 0.5);
-        direction = new Vec3d(0, 123, 0);
+        point = new Vec3f(0.5f, 0.5f, 0.5f);
+        direction = new Vec3f(0, 123, 0);
         assertTrue(quad.containsPoint(point));
         
         // point on plane outside poly
-        point = new Vec3d(-0.5, 0.5, 1.5);
-        direction = new Vec3d(0, 123, 0);
+        point = new Vec3f(-0.5f, 0.5f, 1.5f);
+        direction = new Vec3f(0, 123, 0);
         assertFalse(quad.containsPoint(point));
         
         // point with ray intersecting poly
         assertTrue(quad.intersectsWithRay(0.5f, 0.1f, 0.5f, 0, 1, 0));        
         
         // point with ray intersecting plane outside poly
-        point = new Vec3d(-32, 0.2, 27);
-        direction = new Vec3d(0, 123, 0);
+        point = new Vec3f(-32, 0.2f, 27);
+        direction = new Vec3f(0, 123, 0);
         assertFalse(quad.containsPoint(point));
              
         // point with ray facing away from poly
